fix(preprocesado): double type changed to uint16 to reduce total space

+++ preprocesado.c
@@ -1,6 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <stdint.h>
 
 #define MAX_LINE_LENGTH 100000 
 
@@ -73,7 +74,7 @@ void procesarArchivo(const char *archivoEntrada) {
     fgets(buffer, MAX_LINE_LENGTH, fEntrada);
 
     // Procesar los datos separados por comas
-    double *filaTemp = (double*) malloc(columnas * sizeof(double));
+    uint16_t *filaTemp = (uint16_t*) malloc(columnas * sizeof(uint16_t));
     int contador_filas = 0;
 
     while (fgets(buffer, MAX_LINE_LENGTH, fEntrada) != NULL) {
@@ -94,7 +95,7 @@ void procesarArchivo(const char *archivoEntrada) {
                     filaTemp[col] = 0.0;
                     col++;
                 }
-                fwrite(filaTemp, sizeof(double), columnas, fSalida);
+                fwrite(filaTemp, sizeof(uint16_t), columnas, fSalida);
                 contador_filas++;
             }
         }
